Fix SyntaxWarning problems
Technically, having "error" and "ignore::SyntaxWarning" in the warning
filters should work, but in older Python versions, the "error" elevates
the SyntaxWarning to a SyntaxError and this cannot be undone by another
filter. This works starting in Python 3.12. For now, we make the "error"
filter more specific, so that it does not trigger on the SyntaxWarnings
that we get.